Added contact.py: - inter country matrix creation construct_K - function construct_C,_subdiagonal_array_to_matrix,normalize_matrix moved
